+++ src/test/java/gov/ca/cwds/BaseIntegrationTest.java
@@ -1,6 +1,5 @@
 package gov.ca.cwds;
 
-import static gov.ca.cwds.Constants.H2_DRIVER_CLASS_NAME;
 import static gov.ca.cwds.util.LiquibaseUtils.createCmsDatabase;
 import static gov.ca.cwds.util.LiquibaseUtils.createTokenStoreDatabase;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
@@ -13,6 +12,11 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 public abstract class BaseIntegrationTest {
+
+  public static final String IDM_BASIC_AUTH_USER = "user";
+  public static final String IDM_BASIC_AUTH_PASS = "pass";
+
+  public static final String H2_DRIVER_CLASS_NAME = "org.h2.Driver";
 
   protected MockMvc mockMvc;
